tidy: Fix "multiplication implicit widening" in mythfilerecorder.
Fix the clang-tidy "implicit widening of multiplication result"
warnings.  This check is looking for places where a multiplication
operation may have overflowed the smaller type before the result is
converted to the bigger type. This will give erroneous results.

Convert a variable to the appropriate type.

+++ mythtv/programs/mythfilerecorder/mythfilerecorder.cpp
@@ -93,7 +93,7 @@ void Streamer::SendBytes(void)
         int read_sz = m_blockSize.loadAcquire();
         if (!m_startTime.isValid())
             m_startTime = MythDate::current();
-        int delta = m_startTime.secsTo(MythDate::current()) + 1;
+        qint64 delta = m_startTime.secsTo(MythDate::current()) + 1;
         int rate  = (delta * m_dataRate) - m_dataRead;
 
         read_sz = std::min(rate, read_sz);
